refactor: path aliases
