+++ project-2/hooks/index.ts
@@ -1,1 +1,1 @@
-export { default as useDebounce } from './useDebounce';
+export { default as useDebounce } from './use-debounce';
